% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{bindquestions}
\alias{bindquestions}
\title{Combine multiple variables}
\usage{
bindquestions(data, vnames, keep_var_label = FALSE, keep_val_labels = TRUE)
}
\arguments{
\item{data}{data.frame object}

\item{vnames}{chracter vector of variable names to combine.}

\item{keep_var_label}{Boolean to indicate whether to keep variable label (of the first varaible in \if{html}{\out{<code>}}vnames\if{html}{\out{</code>}}) or not.}

\item{keep_val_labels}{Boolean to indicate whether to keep value labels (of the first varaible in \if{html}{\out{<code>}}vnames\if{html}{\out{</code>}}) or not.}
}
\description{
Combine multiple variables
}
