% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{genfakedata}
\alias{genfakedata}
\title{Generate fake data with random responses using JSON file information}
\usage{
genfakedata(
  survey_json,
  sample_size = 1000,
  rate_missing = 0.01,
  custom_weights = NULL,
  DKtext = "わからない",
  DKcode = 888,
  NAtext = "答えたくない",
  NAcode = 999,
  OtherOptiontext = "その他",
  OtherOptioncode = 666,
  text_randomize = NULL,
  text_starttime = NULL
)
}
\arguments{
\item{survey_json}{Name of JSON survey file exported by gformexperiment custom add-on.}

\item{sample_size}{Sample size for fake data.}

\item{rate_missing}{The rate of missing data to occur for any variable that is not required.}

\item{custom_weights}{Named list object that indicates baseline weights for generating fake answers for designated questions. List element names must correspond to \if{html}{\out{<code>}}name\if{html}{\out{</code>}} and list element lengths must match with the length of \if{html}{\out{<code>}}codes\if{html}{\out{</codes>}} (+1 if there is 'other' option).}

\item{DKtext}{Raw full text(s) of the response option(s) that indicate "don't know" answers.}

\item{DKcode}{Response code for "don't know" answers.}

\item{NAtext}{Raw full text(s) of the response option(s) that indicate refused answers.}

\item{NAcode}{Response code for refused answers.}

\item{OtherOptiontext}{Texts in the codebook to indicate "other" answers.}

\item{OtherOptioncode}{Response code for "other" answers.}

\item{text_randomize}{Raw full text(s) of the question(s) that are used to randomize experimental conditions.}

\item{text_starttime}{Raw full text of the question that is used to record starting time.}
}
\description{
Generate fake data with random responses using JSON file information
}
