% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{gensimplecodebook}
\alias{gensimplecodebook}
\title{Generate (and export) simplified codebook}
\usage{
gensimplecodebook(
  survey_json,
  DKtext = "わからない",
  DKcode = 888,
  NAtext = "答えたくない",
  NAcode = 999,
  OtherOptiontext = "その他",
  OtherOptioncode = 666,
  includeTimeStamp = TRUE,
  includeEmail = FALSE,
  text_randomize = NULL,
  text_starttime = NULL,
  show_duration = TRUE,
  textout = NULL,
  csvout = NULL
)
}
\arguments{
\item{survey_json}{JSON survey file exported through surveyexport.js.}

\item{DKtext}{Raw full text(s) of the response option(s) that indicate "don't know" answers.}

\item{DKcode}{Response code for "don't know" answers.}

\item{NAtext}{Raw full text(s) of the response option(s) that indicate refused answers.}

\item{NAcode}{Response code for refused answers.}

\item{OtherOptiontext}{Texts in the codebook to indicate "other" answers.}

\item{OtherOptioncode}{Response code for "other" answers.}

\item{includeTimeStamp}{Boolean to indicate whether to include system-generated timestamp variable in the codebook or not.}

\item{includeEmail}{Boolean to indicate whether to include system-generated email variable in the codebook or not.}

\item{text_randomize}{Raw full text(s) of the question(s) that are used to randomize experimental conditions.}

\item{text_starttime}{Raw full text of the question that is used to record starting time.}

\item{show_duration}{Boolean to indicate whether to include duration variable in the codebook or not (if \if{html}{\out{<code>}}text_starttime\if{html}{\out{</code>}} is available).}

\item{textout}{Name and path of the output html file to be saved.}

\item{csvout}{Name and path of the output csv file to be saved.}
}
\description{
Generate (and export) simplified codebook
}
