% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{read_gform}
\alias{read_gform}
\title{Import Google Form survey CSV Data using JSON file information}
\usage{
read_gform(
  responses_data,
  survey_json,
  DKtext = "わからない",
  DKcode = 888,
  NAtext = "答えたくない",
  NAcode = 999,
  OtherOptiontext = "その他",
  OtherOptioncode = 666,
  text_randomize = NULL,
  text_starttime = NULL,
  includeTimeStamp = TRUE,
  show_duration = TRUE,
  includeEmail = FALSE,
  survey_csv = NULL,
  ...
)
}
\arguments{
\item{responses_data}{Name of either one of (1) JSON responses file exported by gformexperiment custom add-on or (2) CSV survey response file exported using Google Sheets output.}

\item{survey_json}{Name of JSON survey file exported by gformexperiment custom add-on.}

\item{DKtext}{Raw full text(s) of the response option(s) that indicate "don't know" answers.}

\item{DKcode}{Response code for "don't know" answers.}

\item{NAtext}{Raw full text(s) of the response option(s) that indicate refused answers.}

\item{NAcode}{Response code for refused answers.}

\item{OtherOptiontext}{Texts in the codebook to indicate "other" answers.}

\item{OtherOptioncode}{Response code for "other" answers.}

\item{text_randomize}{Raw full text(s) of the question(s) that are used to randomize experimental conditions.}

\item{text_starttime}{Raw full text of the question that is used to record starting time.}

\item{includeTimeStamp}{Boolean to indicate whether to include system-generated timestamp variable in the codebook or not.}

\item{show_duration}{Boolean to indicate whether to include duration variable in the codebook or not (if \if{html}{\out{<code>}}text_starttime\if{html}{\out{</code>}} is available).}

\item{includeEmail}{Boolean to indicate whether to include system-generated email variable in the codebook or not.}

\item{survey_csv}{(deprecated) import responses data in CSV file (incorporated by \if{html}{\out{<code>}}responses_data\if{html}{\out{</code>}}).}

\item{...}{Optional argumens passed to \if{html}{\out{<code>}}read.csv()\if{html}{\out{</code>}} function.}
}
\description{
Import Google Form survey CSV Data using JSON file information
}
