% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{bindquestions}
\alias{bindquestions}
\title{Combine multiple variables}
\usage{
bindquestions(
  data,
  vnames,
  keep_var_label = FALSE,
  keep_val_labels = TRUE,
  reversed_loc = NULL,
  val_lim = NULL
)
}
\arguments{
\item{data}{data.frame object}

\item{vnames}{character vector of variable names to combine.}

\item{keep_var_label}{Boolean to indicate whether to keep variable label (of the first variable in \if{html}{\out{<code>}}vnames\if{html}{\out{</code>}}) or not.}

\item{keep_val_labels}{Boolean to indicate whether to keep value labels (of the first variable in \if{html}{\out{<code>}}vnames\if{html}{\out{</code>}}) or not.}

\item{reversed_loc}{Integer vector to indicate the location of variables to be reversed in \if{html}{\out{<code>}}vnames\if{html}{\out{</code>}}.}

\item{val_lim}{Integer vector of length 2 that indicates the limits of values (only used when \if{html}{\out{<code>}}reversed_loc\if{html}{\out{</code>}} is not NULL.)}
}
\description{
Combine multiple variables
}
