% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{gencodebook}
\alias{gencodebook}
\title{Generate full-information codebook dataset}
\usage{
gencodebook(
  survey_json,
  DKtext = "わからない",
  DKcode = 888,
  NAtext = "答えたくない",
  NAcode = 999,
  text_randomize = NULL,
  text_starttime = NULL,
  attach_gotoPage = TRUE
)
}
\arguments{
\item{survey_json}{Name of JSON survey file exported by gformexperiment custom add-on.}

\item{DKtext}{Raw full text(s) of the response option(s) that indicate "don't know" answers.}

\item{DKcode}{Response code for "don't know" answers.}

\item{NAtext}{Raw full text(s) of the response option(s) that indicate refused answers.}

\item{NAcode}{Response code for refused answers.}

\item{text_randomize}{Raw full text(s) of the question(s) that are used to randomize experimental conditions.}

\item{text_starttime}{Raw full text of the question that is used to record starting time.}

\item{attach_gotoPage}{If there is page transition conditional on answers, attach the next page to choices.}
}
\description{
Generate full-information codebook dataset
}
