% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{genstartpage}
\alias{genstartpage}
\title{Generate experiment start page (HTML)}
\usage{
genstartpage(survey_json, text_randomize, text_starttime = NULL, file = NULL)
}
\arguments{
\item{survey_json}{JSON survey file exported through surveyexport.js.}

\item{text_randomize}{Raw full text(s) of the question(s) that are used to randomize experimental conditions.}

\item{text_starttime}{Raw full text of the question that is used to record starting time.}

\item{file}{Name and path of the output html file to be saved.}
}
\description{
Generate experiment start page (HTML)
}
