% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformexperiment.R
\name{read_gform}
\alias{read_gform}
\title{Import Google Form survey CSV Data using JSON file information}
\usage{
read_gform(
  survey_csv,
  survey_json,
  DKtext = "\\\\u308f\\\\u304b\\\\u3089\\\\u306a\\\\u3044",
  DKcode = 888,
  NAtext = "\\\\u7b54\\\\u3048\\\\u305f\\\\u304f\\\\u306a\\\\u3044",
  NAcode = 999,
  OtherOptiontext = "\\\\u305d\\\\u306e\\\\u4ed6",
  OtherOptioncode = 666,
  text_randomize = NULL,
  text_starttime = NULL,
  includeTimeStamp = TRUE,
  show_duration = TRUE,
  includeEmail = FALSE,
  ...
)
}
\arguments{
\item{survey_csv}{CSV survey response file exported using Google Sheets output.}

\item{survey_json}{JSON survey file exported through surveyexport.js.}

\item{DKtext}{Raw full text(s) of the response option(s) that indicate "don't know" answers.}

\item{DKcode}{Response code for "don't know" answers.}

\item{NAtext}{Raw full text(s) of the response option(s) that indicate refused answers.}

\item{NAcode}{Response code for refused answers.}

\item{OtherOptiontext}{Texts in the codebook to indicate "other" answers.}

\item{OtherOptioncode}{Response code for "other" answers.}

\item{text_randomize}{Raw full text(s) of the question(s) that are used to randomize experimental conditions.}

\item{text_starttime}{Raw full text of the question that is used to record starting time.}

\item{includeTimeStamp}{Boolean to indicate whether to include system-generated timestamp variable in the codebook or not.}

\item{show_duration}{Boolean to indicate whether to include duration variable in the codebook or not (if \if{html}{\out{<code>}}text_starttime\if{html}{\out{</code>}} is available).}

\item{includeEmail}{Boolean to indicate whether to include system-generated email variable in the codebook or not.}

\item{...}{Optional argumens passed to \if{html}{\out{<code>}}read.csv()\if{html}{\out{</code>}} function.}
}
\description{
Import Google Form survey CSV Data using JSON file information
}
